requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
av==16.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.3.1
flatbuffers==25.12.19
frozenlist==1.8.0
google-api-core==2.30.0
google-auth==2.49.0
google-cloud-speech==2.37.0
google-cloud-texttospeech==2.34.0
google-genai==1.66.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
jiter==0.13.0
livekit==1.1.2
livekit-agents==1.4.4
livekit-api==1.1.0
livekit-blingfire==1.1.0
livekit-plugins-google==1.4.4
livekit-plugins-openai==1.4.4
livekit-plugins-sarvam==1.4.4
livekit-plugins-silero==1.4.4
livekit-protocol==1.1.2
markdown-it-py==4.0.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
nest-asyncio==1.6.0
numpy==2.4.2
onnxruntime==1.24.3
openai==2.26.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1  
opentelemetry-exporter-otlp-proto-grpc==1.39.1    
opentelemetry-exporter-otlp-proto-http==1.39.1    
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1        
packaging==26.0
pillow==12.1.1
prometheus_client==0.24.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pypdf==6.9.0
python-dotenv==1.2.2
requests==2.32.5
rich==14.3.3
rsa==4.9.1
shellingham==1.5.4
sniffio==1.3.1
sounddevice==0.5.5
sympy==1.14.0
tenacity==9.1.4
tqdm==4.67.3
typer==0.24.1
types-protobuf==6.32.1.20260221
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
watchfiles==1.1.1
websockets==15.0.1
yarl==1.23.0
zipp==3.23.0
pymongo==4.6.1
